lombok 도입 및 Repository 인터페이스 제거
