+++ .gitignore
@@ -1,6 +1,5 @@
 HELP.md
 .gradle
-build/
 !gradle/wrapper/gradle-wrapper.jar
 
 ### IntelliJ IDEA ###
