+++ src/main/java/com/example/todolistjava/domain/Todo.java
@@ -1,5 +1,9 @@
 package com.example.todolistjava.domain;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
 public class Todo {
     private String id;
     private String text;
@@ -11,37 +15,5 @@ public class Todo {
         this.text = text;
         this.isCompleted = isCompleted;
         this.isEdit = isEdit;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public boolean isCompleted() {
-        return isCompleted;
-    }
-
-    public void setCompleted(boolean completed) {
-        isCompleted = completed;
-    }
-
-    public boolean isEdit() {
-        return isEdit;
-    }
-
-    public void setEdit(boolean edit) {
-        isEdit = edit;
     }
 }
