+++ src/main/java/com/example/todolistjava/repository/TodoRepository.java
@@ -1,10 +1,24 @@
 package com.example.todolistjava.repository;
 
 import com.example.todolistjava.domain.Todo;
+import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import java.util.*;
 
-public interface TodoRepository {
-    void save(Todo todo);
-    List<Todo> findAll();
+@Repository
+public class TodoRepository {
+
+    List<Todo> store = new ArrayList<>(){{
+        add(new Todo("1", "a", false, false));
+        add(new Todo("2", "b", false, false));
+        add(new Todo("3", "c", false, false));
+    }};
+
+    public void save(Todo todo) {
+        store.add(todo);
+    }
+
+    public List<Todo> findAll() {
+        return store;
+    }
 }
