--- src/test/java/com/example/todolistjava/TodoListJavaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.todolistjava;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TodoListJavaApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
